+++ htp/htp_private.h
@@ -452,8 +452,8 @@ void htp_normalize_uri_path_inplace(bstr *s);
 void htp_utf8_decode_path_inplace(htp_cfg_t *cfg, htp_tx_t *tx, bstr *path);
 void htp_utf8_validate_path(htp_tx_t *tx, bstr *path);
 
-int htp_parse_content_length(bstr *b);
-int htp_parse_chunked_length(unsigned char *data, size_t len);
+int64_t htp_parse_content_length(bstr *b);
+int64_t htp_parse_chunked_length(unsigned char *data, size_t len);
 int64_t htp_parse_positive_integer_whitespace(unsigned char *data, size_t len, int base);
 int htp_parse_status(bstr *status);
 int htp_parse_authorization_digest(htp_connp_t *connp, htp_header_t *auth_header);
